refactor: change event listener with vglobal #3734
